New DD tests, adding dirwalk and filewalk

+++ tsdsfe/test/test.sh
@@ -4,12 +4,18 @@ echo "Starting TSDSFE server."
 node tsdsfe.js --checkservers false --checkdeps false &
 PID=$!
 
+RESULT=$?
+if [[ $RESULT != "0" ]]; then
+	echo "test.sh Could not start TSDSFE server.  Exiting with code 1."
+	exit $RESULT
+fi
+
 echo "Sleeping for 3 seconds before running tests."
 
 sleep 3
 
 rm -rf node_modules/datacache/cache/*
-rm -rf ../cache/*;
+rm -rf ../cache/*
 
 node test/test.js --testfile test/metadata-tests.js
 
@@ -30,9 +36,9 @@ RESULT+=$?
 kill -s "SIGINT" $PID
 
 if [[ $RESULT != "1" ]]; then
-	echo "test.sh Exiting with code 1"
+	echo "test.sh Exiting with code 1."
 	exit 1
 else
-	echo "test.sh Exiting with code 0"
+	echo "test.sh Exiting with code 0."
 	exit 0
 fi
